DT-775 expand filter values in selection list
