--- ui/src/app/pages/data/cohort/variant-filter-selection.tsx
@@ -1,16 +1,0 @@
-import * as React from 'react';
-
-import {
-  CriteriaType,
-  Domain,
-  Variant,
-  VariantFilterRequest,
-} from 'generated/fetch';
-
-
-export const VariantFilterSelection = () => {
-
-  return (
-    <></>
-  );
-}
